Created a global feature context so that new step definitions go to all projects that extend it.
